vao as well as some common files

+++ src/bebone/gfx/bo.cpp
@@ -2,10 +2,10 @@
 
 using namespace bebone::gfx;
 
-unsigned long BO::get_id() const {
+unsigned int BO::get_id() const {
     return _ID; 
 }
 
-unsigned long* BO::get_id_ref() {
+unsigned int* BO::get_id_ref() {
     return &_ID;
 }
